[completed]file storage layer

+++ mora-core/src/models/queues.rs
@@ -15,5 +15,5 @@ pub struct PostQueueRequest {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Queue {
     pub id: String,
-    pub pending_events_count: usize,
+    pub pending_events_count: u128,
 }
